+++ src/slash/8ball.ts
@@ -1,32 +1,37 @@
 import FluorineClient from '@classes/Client';
 import Embed from '@classes/Embed';
-import { Message } from 'discord.js';
+import { CommandInteraction } from 'discord.js';
+import { SlashCommandBuilder } from '@discordjs/builders';
+import { Category } from 'types/applicationCommand';
+import hash from 'murmurhash-v3';
+
 export async function run(
     client: FluorineClient,
-    message: Message,
-    args: string[]
+    interaction: CommandInteraction
 ) {
+    const question = interaction.options.getString('question');
     const responses = client.language.get(
-        message.guild.preferredLocale,
+        interaction.locale,
         '8BALL_RESPONSES'
     );
-    if (!args[0]) {
-        return message.reply(
-            client.language.get(message.guild.preferredLocale, '8BALL_ERROR')
-        );
-    }
 
-    const embed = new Embed(client, message.guild.preferredLocale)
-        .setDescription(args.join(' '))
+    const embed = new Embed(client, interaction.locale)
+        .setDescription(question)
         .addLocaleField({
             name: '8BALL_RESPONSE',
-            value: responses[Math.floor(Math.random() * responses.length)]
+            value: responses[hash(question) % responses.length]
         });
-    message.reply({ embeds: [embed] });
+    interaction.reply({ embeds: [embed] });
 }
-export const help = {
-    name: '8ball',
-    description: 'Magiczna kula',
-    aliases: [],
-    category: 'fun'
-};
+
+export const data = new SlashCommandBuilder()
+    .setName('8ball')
+    .setDescription('Ask the magic ball a question')
+    .addStringOption(option =>
+        option
+            .setName('question')
+            .setDescription('Ask a question')
+            .setRequired(true)
+    );
+
+export const category: Category = 'fun';
