+++ src/slash/ban.ts
@@ -1,99 +1,111 @@
 import FluorineClient from '../classes/Client';
 import Embed from '../classes/Embed';
-import { Message } from 'discord.js';
+import { CommandInteraction } from 'discord.js';
+import { SlashCommandBuilder } from '@discordjs/builders';
 import createCase from '../util/createCase';
 import r from 'rethinkdb';
 import modLog from '@util/modLog';
+import { Category } from 'types/applicationCommand';
+
 export async function run(
     client: FluorineClient,
-    message: Message,
-    args: string[]
+    interaction: CommandInteraction<'cached'>
 ) {
-    if (!message.member?.permissions.has('BAN_MEMBERS')) {
-        return message.reply(
-            client.language.get(
-                message.guild.preferredLocale,
+    if (!interaction.member?.permissions.has('BAN_MEMBERS')) {
+        return interaction.reply({
+            content: client.language.get(
+                interaction.locale,
                 'BAN_PERMISSIONS_MISSING'
-            )
-        );
+            ),
+            ephemeral: true
+        });
     }
 
-    if (!args[0])
-        return message.reply(
-            client.language.get(
-                message.guild.preferredLocale,
-                'BAN_ARGUMENTS_MISSING'
-            )
-        );
-
-    const member =
-        message.mentions.members?.first() ??
-        (await message.guild?.members.fetch(args[0]).catch(() => null));
+    const member = interaction.options.getMember('user');
     const reason =
-        args.slice(1).join(' ') ||
-        client.language.get(message.guild.preferredLocale, 'NONE');
+        interaction.options.getString('reason') ??
+        client.language.get(interaction.locale, 'NONE');
 
     if (!member)
-        return message.reply(
-            client.language.get(
-                message.guild.preferredLocale,
+        return interaction.reply({
+            content: client.language.get(
+                interaction.locale,
                 'BAN_MEMBER_MISSING'
-            )
-        );
+            ),
+            ephemeral: true
+        });
 
-    if (!member?.bannable)
-        return message.reply(
-            client.language.get(
-                message.guild.preferredLocale,
+    if (member.user.id === interaction.user.id)
+        return interaction.reply({
+            content: client.language.get(
+                interaction.locale,
+                'BAN_ERROR_YOURSELF'
+            ),
+            ephemeral: true
+        });
+
+    if (!member.bannable)
+        return interaction.reply({
+            content: client.language.get(
+                interaction.locale,
                 'BAN_BOT_PERMISSIONS_MISSING'
-            )
-        );
+            ),
+            ephemeral: true
+        });
 
     if (reason.length > 1024) {
-        message.reply(
-            client.language.get(
-                message.guild.preferredLocale,
+        return interaction.reply({
+            content: client.language.get(
+                interaction.locale,
                 'REASON_LONGER_THAN_1024'
-            )
-        );
+            ),
+            ephemeral: true
+        });
     }
 
     const create = await createCase(
         client,
-        message?.guild,
+        interaction?.guild,
         member.user,
-        message.author,
+        interaction.user,
         'ban',
         reason
     );
 
-    member.ban({
-        reason: client.language.get(
-            message.guild.preferredLocale,
-            'BAN_REASON',
-            {
-                user: message.author.tag,
-                reason
-            }
-        )
+    await member.ban({
+        reason: client.language.get(interaction.locale, 'BAN_REASON', {
+            user: interaction.user.tag,
+            reason
+        })
     });
-
-    modLog(client, create, message.guild);
-    const embed = new Embed(client, message.guild.preferredLocale)
+    modLog(client, create, interaction.guild);
+    const embed = new Embed(client, interaction.locale)
         .setLocaleTitle('BAN_SUCCESS_TITLE')
         .setLocaleDescription('BAN_SUCCESS_DESCRIPTION')
-        .setThumbnail(member.user.displayAvatarURL({ dynamic: true }))
-        .addLocaleField({ name: 'BAN_MODERATOR', value: message.author.tag })
+        .setThumbnail(member.displayAvatarURL({ dynamic: true }))
+        .addLocaleField({ name: 'BAN_MODERATOR', value: interaction.user.tag })
         .addLocaleField({ name: 'BAN_USER', value: member.user.tag })
         .addLocaleField({ name: 'REASON', value: reason })
         .addLocaleField({ name: 'PUNISHMENT_ID', value: create.id.toString() });
-    message.reply({ embeds: [embed] });
+    interaction.reply({ embeds: [embed] });
 
     r.table('case').insert(create).run(client.conn);
 }
-export const help = {
-    name: 'ban',
-    description: 'Zbanuj kogoś z serwera',
-    aliases: ['zbanuj'],
-    category: 'moderation'
-};
+
+export const data = new SlashCommandBuilder()
+    .setName('ban')
+    .setDescription('Ban an user from the server')
+    .addUserOption(option =>
+        option
+            .setName('user')
+            .setDescription('Provide an user to ban')
+            .setRequired(true)
+    )
+    .addStringOption(option =>
+        option
+            .setName('reason')
+            .setDescription('Provide a reason for banning this user')
+            .setRequired(false)
+    );
+
+export const category: Category = 'moderation';
