+++ src/slash/ping.ts
@@ -1,16 +1,21 @@
 import FluorineClient from '@classes/Client';
 import Embed from '@classes/Embed';
-import { Message } from 'discord.js';
+import { SlashCommandBuilder } from '@discordjs/builders';
+import { CommandInteraction } from 'discord.js';
+import { Category } from 'types/applicationCommand';
 
-export async function run(client: FluorineClient, message: Message) {
-    const embed = new Embed(client, message.guild.preferredLocale)
+export async function run(
+    client: FluorineClient,
+    interaction: CommandInteraction
+) {
+    const embed = new Embed(client, interaction.locale)
         .setTitle('Ping')
         .addLocaleField({ name: 'PING', value: `${client.ws.ping}ms` });
-    message.reply({ embeds: [embed] });
+    interaction.reply({ embeds: [embed] });
 }
-export const help = {
-    name: 'ping',
-    description: 'Opóźnienie pomiędzy discordem a botem',
-    aliases: ['opoznienie', 'pong'],
-    category: 'tools'
-};
+
+export const data = new SlashCommandBuilder()
+    .setName('ping')
+    .setDescription("Check the bot's ping");
+
+export const category: Category = 'tools';
